requirements.txt
# stable version for development
Flask==2.0.1
twilio==8.4.0
requests==2.26.0
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2023.3

# i dont know which one works bruh
dotenv-python==0.0.1
python-dotenv==1.0.0